Migrate the base widget to typescript.
